refactor: 💡 Add Short Path and Handle Code
