+++ apps/rose/src/app/core/enums/sortobject.ts
@@ -1,0 +1,6 @@
+export enum sortObjectVal {
+  P_ASC = "price-asc",
+  P_DESC = "price-desc",
+  A_DESC = "alpha-desc",
+  A_ASC = "alpha-asc",
+}
