+++ apps/rose/src/app/core/enums/sorttypes.ts
@@ -1,0 +1,4 @@
+export enum sortTypes {
+  ASC = "asc",
+  DESC = "desc",
+}
