+++ apps/rose/src/app/features/layouts/dashboard/dashboard.component.ts
@@ -1,8 +1,10 @@
 import { Component } from "@angular/core";
+// Router
 import { RouterOutlet } from "@angular/router";
-
-import { FooterComponent } from "../footer/footer.component";
+// Components
 import { NavbarComponent } from "../navbar/navbar.component";
+import { FooterComponent } from "../footer/footer.component";
+
 
 
 @Component({
@@ -10,6 +12,6 @@ import { NavbarComponent } from "../navbar/navbar.component";
   imports: [RouterOutlet, NavbarComponent, FooterComponent],
   templateUrl: "./dashboard.component.html",
   styleUrl: "./dashboard.component.scss",
-  
+
 })
 export class DashboardComponent {}
