+++ apps/rose/src/app/features/pages/all-categories/all-categories.component.ts
@@ -1,32 +1,30 @@
 import { Component, DestroyRef, inject, OnInit, signal } from "@angular/core";
-
 import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
-
+// Animations
+import { animate, query, stagger, style, transition, trigger } from '@angular/animations';
+// Translation
 import { TranslatePipe } from "@ngx-translate/core";
-
-import { Product } from "../../../core/interfaces/carditem.interface";
-
-import { CardItemComponent } from "../../../shared/components/ui/card-item/card-item.component";
-import { ProductsService } from "../../../shared/services/products/products.service";
-
-import { Store } from "@ngrx/store";
-import { loadProductsToFilter } from "../../../store/filter/filter.actions";
-import { selectFilterProducts } from "../../../store/filter/filter.selector";
-import * as sortActions from "../../../store/sort/sort.actions";
-import * as sortSelectors from "../../../store/sort/store.selectors";
-
+import { TranslationService } from "@rose/core_services/translation/translation.service";
+// Animations_Translation
+import { fadeTransition } from "@rose/core_services/translation/fade.animation";
+// Interfaces
+import { Product } from "@rose/core_interfaces/carditem.interface";
+// Components
 import { FilterCategoriesComponent } from "./components/filter-categories/filter-categories.component";
-
+// Shared_Components
+import { NoDataAvailableComponent } from "@rose/shared_Components_business/no-data-available/no-data-available.component";
+import { CardItemComponent } from "@rose/shared_Components_ui/card-item/card-item.component";
+// Shared_Services
+import { ProductsService } from "@rose/shared_services/products/products.service";
+//NGRX
+import { Store } from "@ngrx/store";
+import { loadProductsToFilter } from "@rose/store_filter/filter.actions";
+import { selectFilterProducts } from "@rose/store_filter/filter.selector";
+import * as sortActions from "@rose/store_sort/sort.actions";
+import * as sortSelectors from "@rose/store_sort/store.selectors";
 // primeng
 import { ButtonModule } from "primeng/button";
 import { DrawerModule } from "primeng/drawer";
-
-import { animate, query, stagger, style, transition, trigger } from '@angular/animations';
-import { NoDataAvailableComponent } from "../../../shared/components/business/no-data-available/no-data-available.component";
-import { TranslationService } from "../../../core/services/translation/translation.service";
-import { fadeTransition } from "../../../core/services/translation/fade.animation";
-
-
 
 @Component({
   selector: "app-all-categories",
