+++ apps/rose/src/app/features/pages/all-categories/components/filter-categories/components/filter-category/filter-category.component.ts
@@ -1,17 +1,18 @@
 import { Component, DestroyRef, inject, OnInit } from "@angular/core";
 import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
-
+// Translation
 import { TranslatePipe } from "@ngx-translate/core";
-
+// NGRX
 import { Store } from "@ngrx/store";
-import { loadSelectedCategories } from './../../../../../../../store/filter/filter.actions';
-
-import { CheckedCardComponent } from "../../../../../../../shared/components/business/checkbox/checked-card.component";
-import { FilterCardComponent } from "../../../../../../../shared/components/ui/filter-card/filter-card.component";
-import { ProductsService } from "../../../../../../../shared/services/products/products.service";
-
-import { CategoryProductCount } from "../../../../../../../core/interfaces/count-by-product.interface";
-import { selectedItem } from "./../../../../../../../core/interfaces/filter-item.interface";
+import { loadSelectedCategories } from "@rose/store_filter/filter.actions";
+// Interfaces
+import { CategoryProductCount } from "@rose/core_interfaces/count-by-product.interface";
+import { selectedItem } from "@rose/core_interfaces/filter-item.interface";
+// Shared_Components
+import { FilterCardComponent } from "@rose/shared_Components_ui/filter-card/filter-card.component";
+import { CheckedCardComponent } from "@rose/shared_Components_business/checkbox/checked-card.component";
+// Shared_Services
+import { ProductsService } from "@rose/shared_services/products/products.service";
 
 @Component({
   selector: "app-filter-category",
