+++ apps/rose/src/app/features/pages/all-categories/components/filter-categories/components/filter-price/filter-price.component.ts
@@ -1,15 +1,15 @@
 import { Component, inject } from "@angular/core";
 import { FormsModule } from '@angular/forms';
-
+// Translation
 import { TranslatePipe } from "@ngx-translate/core";
-
-import { FilterCardComponent } from "../../../../../../../shared/components/ui/filter-card/filter-card.component";
-
+// Shared_Components
+import { FilterCardComponent } from "@rose/shared_Components_ui/filter-card/filter-card.component";
+// NGRX
 import { Store } from "@ngrx/store";
-import { loadSelectedPrice } from "./../../../../../../../store/filter/filter.actions";
-
+import { loadSelectedPrice } from "@rose/store_filter/filter.actions";
 //Primeng
 import { Slider } from 'primeng/slider';
+
 @Component({
   selector: "app-filter-price",
   imports: [FilterCardComponent, TranslatePipe, FormsModule, Slider],
