+++ apps/rose/src/app/features/pages/all-categories/components/filter-categories/components/filter-rating/filter-rating.component.ts
@@ -1,16 +1,14 @@
 import { Component, inject } from "@angular/core";
-
+// Translation
 import { TranslatePipe } from "@ngx-translate/core";
-
-import { FilterCardComponent } from "../../../../../../../shared/components/ui/filter-card/filter-card.component";
-import { CheckedCardComponent } from "../../../../../../../shared/components/business/checkbox/checked-card.component";
-import {
-  FilterItem,
-  selectedItem,
-} from "./../../../../../../../core/interfaces/filter-item.interface";
-
+// Interfaces
+import { FilterItem, selectedItem } from "@rose/core_interfaces/filter-item.interface";
+// Shared_Components
+import { FilterCardComponent } from "@rose/shared_Components_ui/filter-card/filter-card.component";
+import { CheckedCardComponent } from "@rose/shared_Components_business/checkbox/checked-card.component";
+// NGRX
 import { Store } from "@ngrx/store";
-import { loadSelectedRating } from '../../../../../../../store/filter/filter.actions';
+import { loadSelectedRating } from '@rose/store_filter/filter.actions';
 
 @Component({
   selector: "app-filter-rating",
