+++ apps/rose/src/app/features/pages/all-categories/components/filter-categories/components/sorting/sorting.component.ts
@@ -1,18 +1,14 @@
 import { Component, inject } from "@angular/core";
-
+// Translation
 import { TranslatePipe } from "@ngx-translate/core";
-
-import { FilterCardComponent } from "../../../../../../../shared/components/ui/filter-card/filter-card.component";
-
+// Shared_Components
+import { FilterCardComponent } from "@rose/shared_Components_ui/filter-card/filter-card.component";
+// NGRX
 import { Store } from "@ngrx/store";
-import * as sortActions from "../../../../../../../store/sort/sort.actions";
-import { sortType } from "../../../../../../../store/sort/sort.states";
-export enum sortObjectVal {
-  P_ASC = "price-asc",
-  P_DESC = "price-desc",
-  A_DESC = "alpha-desc",
-  A_ASC = "alpha-asc",
-}
+import * as sortActions from "@rose/store_sort/sort.actions";
+import { sortType } from "@rose/store_sort/sort.states";
+// Enums
+import { sortObjectVal } from "@rose/core_enums/sortobject";
 @Component({
   selector: "app-sorting",
   imports: [TranslatePipe, FilterCardComponent],
