+++ apps/rose/src/app/features/pages/all-categories/components/filter-categories/filter-categories.component.ts
@@ -1,21 +1,21 @@
 import { Component, inject } from "@angular/core";
-
+// Translation
 import { TranslatePipe } from "@ngx-translate/core";
-
-import { InputBtnComponent } from "../../../../../shared/components/ui/input-btn/input-btn.component";
-import { FilterCardComponent } from "../../../../../shared/components/ui/filter-card/filter-card.component";
-
+// Directives
+import { EnglishOnlyDirective } from "@rose/core_directive/english-only.directive";
+// Components
 import { SortingComponent } from "./components/sorting/sorting.component";
 import { FilterOccasionsComponent } from "./components/filter-occasions/filter-occasions.component";
 import { FilterPriceComponent } from "./components/filter-price/filter-price.component";
 import { FilterSalesComponent } from "./components/filter-sales/filter-sales.component";
 import { FilterRatingComponent } from "./components/filter-rating/filter-rating.component";
 import { FilterCategoryComponent } from "./components/filter-category/filter-category.component";
-
-import { EnglishOnlyDirective } from "../../../../../core/directive/english-only.directive";
-
+// Shared_Components
+import { InputBtnComponent } from "@rose/shared_Components_ui/input-btn/input-btn.component";
+import { FilterCardComponent } from "@rose/shared_Components_ui/filter-card/filter-card.component";
+// NGRX
 import { Store } from "@ngrx/store";
-import { clearFilter, loadSelectedName } from '../../../../../store/filter/filter.actions';
+import { clearFilter, loadSelectedName } from '@rose/store_filter/filter.actions';
 
 @Component({
   selector: "app-filter-categories",
