+++ apps/rose/src/app/features/pages/contact/contact.component.ts
@@ -1,9 +1,8 @@
 import { Component } from "@angular/core";
-import { CommonModule } from "@angular/common";
 
 @Component({
   selector: "app-contact",
-  imports: [CommonModule],
+  imports: [],
   templateUrl: "./contact.component.html",
   styleUrl: "./contact.component.scss",
 })
