+++ apps/rose/src/app/features/pages/home/components/best-seller/bestSeller.component.ts
@@ -1,14 +1,14 @@
 import { Component, inject, OnInit, signal, DestroyRef } from "@angular/core";
 import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
-
+// Translation
 import { TranslatePipe } from "@ngx-translate/core";
-
-import { CardItemComponent } from "../../../../../shared/components/ui/card-item/card-item.component";
-import { ButtonComponent } from "../../../../../shared/components/ui/button/button.component";
-import { BestSellerService } from "../../../../../shared/services/best-seller/best-seller.service";
-
-import { BestSeller, BestSellerRes } from "../../../../../core/interfaces/best-seller.interface";
-
+// Interfaces
+import { BestSeller, BestSellerRes } from "@rose/core_interfaces/best-seller.interface";
+// Shared_Components
+import { CardItemComponent } from "@rose/shared_Components_ui/card-item/card-item.component";
+import { ButtonComponent } from "@rose/shared_Components_ui/button/button.component";
+// Shared_Services
+import { BestSellerService } from "@rose/shared_services/best-seller/best-seller.service";
 //PrimeNg
 import { CarouselModule } from "primeng/carousel";
 import { ButtonModule } from "primeng/button";
