+++ apps/rose/src/app/features/pages/home/home.component.ts
@@ -1,4 +1,9 @@
 import { Component, inject } from "@angular/core";
+// Translation
+import { TranslationService } from "@rose/core_services/translation/translation.service";
+// Animations_Translation
+import { fadeTransition } from "@rose/core_services/translation/fade.animation";
+// Components
 import { CategoriesComponent } from "./components/categories/categories.component";
 import { GiftsComponent } from "./components/gifts/gifts.component";
 import { OurServicesComponent } from "./components/our-services/ourServices.component";
@@ -8,8 +13,6 @@ import { AboutUsComponent } from "./components/about-us/aboutUs.component";
 import { GalleryComponent } from "./components/gallery/gallery.component";
 import { TestimonialsComponent } from "./components/testimonials/testimonials.component";
 import { TrustedByComponent } from "./components/trusted-by/trustedBy.component";
-import { fadeTransition } from "../../../core/services/translation/fade.animation";
-import { TranslationService } from "../../../core/services/translation/translation.service";
 
 @Component({
   selector: "app-home",
