+++ apps/rose/src/app/shared/components/business/translate-toggle/translate-toggle.component.ts
@@ -1,6 +1,6 @@
 import { Component, inject, OnInit, signal, WritableSignal } from "@angular/core";
-import { TranslationService } from "./../../../../core/services/translation/translation.service";
-
+// Translation
+import { TranslationService } from "@rose/core_services/translation/translation.service";
 @Component({
   selector: "app-translate-toggle",
   imports: [ ],
