+++ apps/rose/src/app/shared/components/ui/button-theme/button-theme.component.ts
@@ -1,6 +1,6 @@
 import { Component, inject } from "@angular/core";
-import { DarkModeService } from "../../../../core/services/darkmode/darkmode.service";
-
+// Services
+import { DarkModeService } from "@rose/core_services/darkmode/darkmode.service";
 //PrimeNg
 import { ButtonModule } from "primeng/button";
 import { TooltipModule } from "primeng/tooltip";
