+++ apps/rose/src/app/shared/components/ui/card-item/card-item.component.html
@@ -3,13 +3,7 @@
     @if (loading) {
     <p-skeleton width="100%" height="200px"></p-skeleton>
     } @else if (productInfo) {
-    <img
-      [src]="productInfo.imgCover"
-      [alt]="productInfo.title"
-      class="product-image"
-      loading="lazy"
-      decoding="async"
-    />
+    <img [src]="productInfo.imgCover" [alt]="productInfo.title" class="product-image" loading="lazy" decoding="async" />
     <div class="bg-shape"></div>
     <div class="overlay-icons">
       <i class="pi pi-eye icon-circle"></i>
