+++ apps/rose/src/app/shared/components/ui/search-modal/search-modal.component.ts
@@ -1,5 +1,4 @@
 import { Component, OnInit } from "@angular/core";
-
 //primeNg
 import { ButtonModule } from "primeng/button";
