+++ apps/rose/src/app/shared/services/best-seller/best-seller.service.ts
@@ -1,8 +1,10 @@
 import { Injectable } from "@angular/core";
 import { HttpClient } from "@angular/common/http";
 import { Observable } from "rxjs";
-import { BestSellerRes } from "../../../core/interfaces/best-seller.interface";
-import { EndPoint } from "../../../core/enums/endpoints";
+// Interfaces
+import { BestSellerRes } from "@rose/core_interfaces/best-seller.interface";
+// Enums
+import { EndPoint } from "@rose/core_enums/endpoints";
 
 @Injectable({
   providedIn: "root",
