+++ apps/rose/src/app/store/filter/filter.actions.ts
@@ -1,8 +1,7 @@
 import { createAction, props } from "@ngrx/store";
-import { Product } from "../../core/interfaces/carditem.interface";
-import { selectedItem } from "../../core/interfaces/filter-item.interface";
-
-4;
+// Interfaces
+import { Product } from "@rose/core_interfaces/carditem.interface";
+import { selectedItem } from "@rose/core_interfaces/filter-item.interface";
 
 export const loadProductsToFilter = createAction(
   "[FILTER] Load Products",
