+++ apps/rose/src/app/store/filter/filter.effect.ts
@@ -1,8 +1,8 @@
 import { inject } from "@angular/core";
-import { Actions, createEffect, ofType } from "@ngrx/effects";
 import { Store } from "@ngrx/store";
-import { ApplyFilters, loadProductsToFilter, loadSelectedCategories, loadSelectedOccasions, loadSelectedPrice, loadSelectedName, loadSelectedRating } from "./filter.actions";
+import { Actions, createEffect, ofType } from "@ngrx/effects";
 import { tap } from "rxjs/operators";
+import { ApplyFilters, loadProductsToFilter, loadSelectedCategories, loadSelectedOccasions, loadSelectedPrice, loadSelectedName, loadSelectedRating } from "./filter.actions";
 
 export class FilterEffects {
   private readonly _actions$ = inject(Actions);
