+++ apps/rose/src/app/store/sort/sort.states.ts
@@ -1,8 +1,7 @@
-import { Product } from "../../core/interfaces/carditem.interface";
+// Interface
+import { Product } from "@rose/core_interfaces/carditem.interface";
 
 export type sortType = "asc" | "desc";
-
-
 
 export interface sortConditions {
   sortBy: string;
